pyproject.toml
[project]
name = "bill-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.62.0",
    "matplotlib>=3.10.7",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
